pyproject.toml
[tool.poetry]
name = "np-session"
version = "0.0.1"
description = ""
authors = ["Ben Hardcastle <ben.hardcastle@alleninstitute.org>"]
readme = "README.md"
packages = [{include = "np_session", from = "src"}]

[tool.poetry.dependencies]
python = "^3.7.0"
typing-extensions = "*"
psycopg2 = "^2.9.5"
requests = "*"
np_config = "*"
"backports.cached-property" = "*"

[tool.poetry.group.dev.dependencies]
pip-tools = "*"
isort = "*"
mypy = "*"
black = "*"
pytest = ">5.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
    "--doctest-modules",
    # "--pdb",
    # "--pdbcls=IPython.terminal.debugger:TerminalPdb",
]

doctest_optionflags = [
    "NORMALIZE_WHITESPACE",
    "IGNORE_EXCEPTION_DETAIL",
    "ELLIPSIS",
]

testpaths = [
    "tests",
    "src",
]